[ALPS08539520] vcp: update test ipi struct
Correct TEST IPI size to 3 words

MTK-Commit-Id: 1e678c07a626918a29b8f51fcf7def27b1af9793

CR-Id: ALPS08539520
Feature: Video Driver
(cherry-pick from ae6f0eeff9615b3accb3b7ffcc305342989d3a67)

+++ drivers/misc/mediatek/vcp/rv/vcp_helper.h
@@ -152,7 +152,7 @@ struct vcp_regs {
 };
 
 struct vcp_ipi_profile {
-	u8 type;
+	unsigned int type;
 	unsigned int ipi_time_h;
 	unsigned int ipi_time_l;
 };
